[fix] using void return type for set functions and correct size of bitwise operands (#307)
* using void instead of auto return type for set functions with no return

* correct sized bitwise operand

+++ sources/c_api/dml_batch.cpp
@@ -135,7 +135,7 @@ extern "C" dml_status_t dml_batch_set_mem_move_by_index(dml_job_t            *dm
         return status;
     }
     if(flags & DML_FLAG_COPY_ONLY){
-        flags &= ~DML_FLAG_COPY_ONLY;
+        flags &= ~((uint64_t)DML_FLAG_COPY_ONLY);
     }
 
     auto batch = dml::batch(dml_job_ptr->destination_first_ptr, task_count);
